+++ docs/changelog.md
@@ -5,6 +5,7 @@ title: Changelog
 # v0.3.0
 
 * Rename module to pijamas.
+* .empty for arrays, associative arrays and strings.
 
 # v0.2.2